drop->tftype = "init"
