+++ excuter/op-mem-cuda/src/deepx/tf/elementwise_basic.hpp
@@ -1034,73 +1034,7 @@ namespace deepx::tf
         }
     };
 
-    // dropout
-    template <typename Author>
-    class Dropout : public TF
-    {
-    public:
-        Dropout(const vector<Param> &args, const vector<Param> &returns)
-        {
-            this->name = "dropout";
-            this->metadata.author = Author::name();
-            this->tftype = "elementwise";
-            this->args = args;
-            this->returns = returns;
-        }
-        string math_formula() const override
-        {
-            return "T1.dropout(p,seed)->T3";
-        }
-        shared_ptr<TF> clone() const override
-        {
-            return make_shared<Dropout<Author>>(*this);
-        }
-        int run(shared_ptr<MemBase> mem, string &error) override
-        {
-            if (!checktensors({this->args[0].textvalue, this->returns[0].textvalue}, mem, error))
-            {
-                return 1;
-            }
-            Precision a_type = mem->gettensor(this->args[0].textvalue).get()->shape.dtype;
-            Precision c_type = mem->gettensor(this->returns[0].textvalue).get()->shape.dtype;
-            if (a_type != c_type)
-            {
-                error = "Type mismatch: " + precision_str(a_type) + " != " + precision_str(c_type);
-                return 1;
-            }
-            switch (a_type)
-            {
-            case Precision::Float64:
-                tensorfunc::dropout<Author>(*mem->gettensor<double>(this->args[0].textvalue), this->getvar<float>(1, mem), this->getvar<unsigned int>(2, mem), *mem->gettensor<double>(this->returns[0].textvalue));
-                break;
-            case Precision::Float32:
-                tensorfunc::dropout<Author>(*mem->gettensor<float>(this->args[0].textvalue), this->getvar<float>(1, mem), this->getvar<unsigned int>(2, mem), *mem->gettensor<float>(this->returns[0].textvalue));
-                break;
-            case Precision::Float16:
-                tensorfunc::dropout<Author>(*mem->gettensor<half>(this->args[0].textvalue), this->getvar<float>(1, mem), this->getvar<unsigned int>(2, mem), *mem->gettensor<half>(this->returns[0].textvalue));
-                break;
-            case Precision::BFloat16:
-                tensorfunc::dropout<Author>(*mem->gettensor<nv_bfloat16>(this->args[0].textvalue), this->getvar<float>(1, mem), this->getvar<unsigned int>(2, mem), *mem->gettensor<nv_bfloat16>(this->returns[0].textvalue));
-                break;  
-            case Precision::Int64:
-                tensorfunc::dropout<Author>(*mem->gettensor<int64_t>(this->args[0].textvalue), this->getvar<float>(1, mem), this->getvar<unsigned int>(2, mem), *mem->gettensor<int64_t>(this->returns[0].textvalue));
-                break;
-            case Precision::Int32:
-                tensorfunc::dropout<Author>(*mem->gettensor<int32_t>(this->args[0].textvalue), this->getvar<float>(1, mem), this->getvar<unsigned int>(2, mem), *mem->gettensor<int32_t>(this->returns[0].textvalue));
-                break;
-            case Precision::Int16:
-                tensorfunc::dropout<Author>(*mem->gettensor<int16_t>(this->args[0].textvalue), this->getvar<float>(1, mem), this->getvar<unsigned int>(2, mem), *mem->gettensor<int16_t>(this->returns[0].textvalue));
-                break;
-            case Precision::Int8:
-                tensorfunc::dropout<Author>(*mem->gettensor<int8_t>(this->args[0].textvalue), this->getvar<float>(1, mem), this->getvar<unsigned int>(2, mem), *mem->gettensor<int8_t>(this->returns[0].textvalue));
-                break;
-            default:
-                error = "Unsupported dtype: " + precision_str(a_type);
-                return 1;
-            }
-            return 0;
-        }
-    };
+
 };
 
 #endif // DEEPX_TF_ELEMENTWISE_BASIC_HPP
