+++ excuter/op-mem-cuda/src/deepx/tf/init.hpp
@@ -10,7 +10,8 @@
 #include "stdutil/num.hpp"
 namespace deepx::tf
 {
-
+    // 填充
+    // constant
     template <typename Author>
     class Constant : public TF
     {
@@ -90,6 +91,71 @@ namespace deepx::tf
         };
     };
 
+    // dropout
+    template <typename Author>
+    class Dropout : public TF
+    {
+    public:
+        Dropout(const vector<Param> &args, const vector<Param> &returns)
+        {
+            this->name = "dropout";
+            this->metadata.author = Author::name();
+            this->tftype = "init";
+            this->args = args;
+            this->returns = returns;
+        }
+        string math_formula() const override
+        {
+            return "dropout(p,seed)->A";
+        }
+        shared_ptr<TF> clone() const override
+        {
+            return make_shared<Dropout<Author>>(*this);
+        }
+        int run(shared_ptr<MemBase> mem, string &error) override
+        {
+            if (!checktensors({this->returns[0].textvalue }, mem, error))
+            {
+                return 1;
+            }
+            Precision a_type = mem->gettensor(this->returns[0].textvalue).get()->shape.dtype;
+ 
+            switch (a_type)
+            {
+            case Precision::Float64:
+                tensorfunc::dropout<Author>(*mem->gettensor<double>(this->returns[0].textvalue), this->getvar<float>(0, mem), this->getvar<unsigned int>(1, mem));
+                break;
+            case Precision::Float32:
+                tensorfunc::dropout<Author>(*mem->gettensor<float>(this->returns[0].textvalue), this->getvar<float>(0, mem), this->getvar<unsigned int>(1, mem));
+                break;
+            case Precision::Float16:
+                tensorfunc::dropout<Author>(*mem->gettensor<half>(this->returns[0].textvalue), this->getvar<float>(0, mem), this->getvar<unsigned int>(1, mem));
+                break;
+            case Precision::BFloat16:
+                tensorfunc::dropout<Author>(*mem->gettensor<nv_bfloat16>(this->returns[0].textvalue), this->getvar<float>(0, mem), this->getvar<unsigned int>(1, mem));
+                break;
+            case Precision::Int64:
+                tensorfunc::dropout<Author>(*mem->gettensor<int64_t>(this->returns[0].textvalue), this->getvar<float>(0, mem), this->getvar<unsigned int>(1, mem));
+                break;
+            case Precision::Int32:
+                tensorfunc::dropout<Author>(*mem->gettensor<int32_t>(this->returns[0].textvalue), this->getvar<float>(0, mem), this->getvar<unsigned int>(1, mem));
+                break;
+            case Precision::Int16:
+                tensorfunc::dropout<Author>(*mem->gettensor<int16_t>(this->returns[0].textvalue), this->getvar<float>(0, mem), this->getvar<unsigned int>(1, mem));
+                break;
+            case Precision::Int8:
+                tensorfunc::dropout<Author>(*mem->gettensor<int8_t>(this->returns[0].textvalue), this->getvar<float>(0, mem), this->getvar<unsigned int>(1, mem));
+                break;
+            default:
+                error = "Unsupported dtype: " + precision_str(a_type);
+                return 1;
+            }
+            return 0;
+        }
+    };
+
+    // 初始化
+    // arange
     template <typename Author>
     class Arange : public TF
     {
@@ -173,7 +239,6 @@ namespace deepx::tf
             }
             return 0;
         }
-        
     };
 
     template <typename Author>
@@ -201,8 +266,8 @@ namespace deepx::tf
             string name = this->returns[0].textvalue;
             auto tensor = mem->gettensor(name).get();
             auto type = tensor->shape.dtype;
-            int low_pos=0;
-            int high_pos=1;
+            int low_pos = 0;
+            int high_pos = 1;
             unsigned int seed = static_cast<unsigned int>(this->getvar<int>(2, mem));
             switch (type)
             {
@@ -262,7 +327,6 @@ namespace deepx::tf
             }
             return 0;
         }
-        
     };
 
     //
@@ -292,8 +356,8 @@ namespace deepx::tf
             string name = this->returns[0].textvalue;
             auto tensor = mem->gettensor(name).get();
             auto type = tensor->shape.dtype;
-            int mean_pos=0;
-            int stddev_pos=1;
+            int mean_pos = 0;
+            int stddev_pos = 1;
             unsigned int seed = static_cast<unsigned int>(this->getvar<int>(2, mem));
             switch (type)
             {
