+++ excuter/op-mem-ompsimd/src/deepx/tf/init.hpp
@@ -6,7 +6,8 @@
 #include "deepx/tensorfunc/init_miaobyte.hpp"
 #include "stdutil/num.hpp"
 namespace deepx::tf
-{
+{   
+    //填充
     // constant
     template <typename Author>
     class Constant : public TF
@@ -87,6 +88,61 @@ namespace deepx::tf
         };
     };
 
+   //dropout
+    template <typename Author>
+    class Dropout : public TF
+    {
+    public:
+        Dropout(const vector<Param> &args, const vector<Param> &returns)
+        {
+            this->name = "dropout";
+            this->metadata.author = Author::name();
+            this->tftype = "init";
+            this->args = args;
+            this->returns = returns;
+        }
+        string math_formula() const override
+        {
+            return "dropout(p,seed)->A";
+        }
+        shared_ptr<TF> clone() const override
+        {
+            return make_shared<Dropout<Author>>(*this);
+        }
+        int run(shared_ptr<MemBase> mem, string &error) override
+        {
+            Precision a_type = mem->gettensor(this->returns[0].textvalue).get()->shape.dtype;
+    
+            switch (a_type)
+            {
+            case Precision::Float64:
+                tensorfunc::dropout<Author, double>(*mem->gettensor<double>(this->returns[0].textvalue), this->getvar<float>(0,mem,true), this->getvar<unsigned int>(1,mem,true));
+                break;  
+            case Precision::Float32:
+                tensorfunc::dropout<Author, float>(*mem->gettensor<float>(this->returns[0].textvalue), this->getvar<float>(0,mem,true), this->getvar<unsigned int>(1,mem,true));
+                break;
+            case Precision::Int64:
+                tensorfunc::dropout<Author, int64_t>(*mem->gettensor<int64_t>(this->returns[0].textvalue), this->getvar<float>(0,mem,true), this->getvar<unsigned int>(1,mem,true));
+                break;  
+            case Precision::Int32:
+                tensorfunc::dropout<Author, int32_t>(*mem->gettensor<int32_t>(this->returns[0].textvalue), this->getvar<float>(0,mem,true), this->getvar<unsigned int>(1,mem,true));
+                break;
+            case Precision::Int16:
+                tensorfunc::dropout<Author, int16_t>(*mem->gettensor<int16_t>(this->returns[0].textvalue), this->getvar<float>(0,mem,true), this->getvar<unsigned int>(1,mem,true));
+                break;  
+            case Precision::Int8:
+                tensorfunc::dropout<Author, int8_t>(*mem->gettensor<int8_t>(this->returns[0].textvalue), this->getvar<float>(0,mem,true), this->getvar<unsigned int>(1,mem,true));
+                break;
+            default:
+                error = "Unsupported dtype: " + precision_str(a_type);
+                return 1;   
+            }
+            return 0;
+        }
+    };
+
+
+    //初始化
     // arange
     template <typename Author>
     class Arange : public TF
